refactor(dto): standartized dto names
